Assign3 Android App to Manage Movie Descriptions
You are to construct a multiple view Android app for managing
MovieDescriptions. The app should be based on the first single view
Android app that you developed, which used a Json description of a
movie description, and displayed the information in your
MovieDescription class. This app is similar to that one, but uses more
complex controls and multiple views to display and manage a collection
of MovieDescription objects. Add a MovieLibrary class which retains
movie description objects. This app does not need to use the network to
contact a movie database, although a future revision of the app will
use the Open Movie Database API. Your app should provide a list of all
movie titles contained in the collection. It should allow the user to
select one title and see all of the information about that title, and
it should provide the ability to add or remove movie description
entries. New descriptions should be added to the collection by the user
manually entering the appropriate information needed to create a
MovieDescription object. Include at least the following fields: Title,
Year, Rated, Released, Runtime, Genre, Actors, and Plot.