assign5&6 JsonRPC Android&iOS App and Movie Collection Server
You are to create a JsonRPC server in Java that remotely manages a
collection of movie descriptions. You may model the server after the
StudentCollections JsonRPC server that has been provided in the
example: studentCollectionJsonRPC.jar. Refer to the readme.txt file in
this archive for instructions for executing this example. The server
has also been executed as an example in class. Your server should
maintain the same file and directory structure as this example.

In addition to the Movie Collection JsonRPC server, you should create
an Android App, based on your solution to Lab3: Android App to View and
Manage MovieDescriptions. Your Android app should use the server you
create as its only data source for obtaining movie information. You
should not retain a local mirror of the movie collection within the
app, and instead use the JsonRPC methods defined by your collection to
obtain model information. You are free to use either the Thread and
Handler approach or AsyncTask approach as discussed in class to create
your solution. The same packaging, documentation and solution
formatting constraints as with the prior assignments apply to this
assignment as well. These apply to the server you will generate as well
as to the Android app
You are to use the JsonRPC movie server that you created to solve the
last assigment, and construct and iOS app that utilizes the server. The
server manages a collection of movie descriptions.

The iOS app you create to complete this problem should be compatible
with the Android App from the last assignment. That is, your iOS app
running on the simulator and your Android app running on the emulator
should be able to access and jointly utilize your movie server. Both of
your apps should use the server as its only data source for obtaining
movie information. You should not retain a complete local mirror of the
movie collection within the app, and instead use the JsonRPC methods
defined by your collection to obtain model information, and only store
minimal movie information within the app. You may use the iOS AsyncTask
( asyncTaskiOS.playground.jar) discussed in class, or the approach used
in the StudentCollectioniOSJsonRPC sample app as the basis for
connecting to your server.

As discussed in class, to store and access the URL of the server use
the following conventions that will support ease of grading and good
practice. In the iOS app, define a key, value pair in the info.plist
file that provides the URL. The machine address should be the
localhost. So, your default url for the the iOS app may be:

http://localhost:8080/
The simulator runs as a normal process on OSX. Access the value of this
property from the sourcecode with code, assuming the key is
ServerURLString, as follows:
        if let infoPlist = NSBundle.mainBundle().infoDictionary {
            self.urlString = ((infoPlist["ServerURLString"]) as?
String!)!
            NSLog("The default urlString from info.plist is
\(self.urlString)")
        }else{
            NSLog("error getting urlString from info.plist")
        }
The Android app should use an entry in the file: res/values/strings.xml
to record the default URL for android. Since the Android Emulator runs
as a virtual machine its IP is not the same as the host/development
machine. Instead, your development machine is considered as the
Emulator's gateway. To find this on a device or the emulator, use the
Android Debugger from the command line as:
     path_to_sdk/sdk/platform-tools/adb shell
This will remotely log you onto the Emulator. Then enter the following
commands to find ip's:
     ip addr show
     ip route show
Use the default gateway ip, which is: 10.0.2.2. So, your default url
for the the Android app will be:

http://10.0.2.2:8080/