assign 9
Recast both your iOS and Android Movie Library apps: iOS Core Data and
Android Sqlite3 to work with the instructor provided jsonRPC server and
media streamer. The new versions of your apps should use their internal
database and add to that database any movies that are available through
the JsonRPC server. Movies that are available throught the RPC server
will have an additional Json attribute/key which indicates the filename
of the media file. Doing an http request to the instructor supplied
streamer with the filename will cause the file to be streamed to the
Media Player of the mobile device. Your app should use the platform
suppled media player to play in a new scene/view the media file. Thus
your apps will be able to search the OMDB API, add new movie
descriptions, and obtain/save JsonRPC movie descriptions, display movie
information, and where available, play movies for which the streamer
has a media file. Your solution to this problem should be an Xcode
project, and an Android Studio project that works with the instuctor
supplied jsonRPC server and media streamer.

Use the example instructor provided jsonRPC server and media streamer:
mediaPlayerWeb.jar for developing your apps. When extracted, this jar's
project file contains three sub-folders one each for the
streamer/jsonrpcserver, a sample iOS app, and a sample Android app. The
mobile apps demonstrate video playback on each platform by using the
streamer. There is a readme.txt file in the streamer/jsonrpcserver's
folder (MovieLibraryJsonRPCServer) which shows how to run the two
servers.